git-svn-id: file:///svn/branches/AMP_1_14_RELEASE/amp@12177 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ WEB-INF/src/org/dgfoundation/amp/ar/workers/AmountColWorker.java
@@ -3,7 +3,6 @@
  */
 package org.dgfoundation.amp.ar.workers;
 
-import java.math.BigDecimal;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
@@ -64,7 +63,7 @@ public class AmountColWorker extends ColumnWorker {
 	protected Cell getCellFromRow(ResultSet rs) throws SQLException {
 		long ownerId=rs.getLong(1);
 		long objectId=rs.getLong(2);
-		BigDecimal amount = rs.getBigDecimal(3);
+		double amount = rs.getDouble(3);
 		String currency=rs.getString(4);
 		java.sql.Date date= rs.getDate(5);
