+++ WEB-INF/src/org/dgfoundation/amp/harvest/FreeBalanceHarvester.java
@@ -9,7 +9,6 @@ package org.dgfoundation.amp.harvest;
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.LineNumberReader;
-import java.math.BigDecimal;
 import java.text.DecimalFormat;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -299,7 +298,7 @@ public class FreeBalanceHarvester {
 						if(element.getTransactionAmount()!=null) oldVal=element.getTransactionAmount().doubleValue();
 						double newVal=n.doubleValue();
 						
-						element.setTransactionAmount(new BigDecimal(oldVal+newVal));
+						element.setTransactionAmount(new Double(oldVal+newVal));
 						
 						break;
 					}
@@ -315,7 +314,7 @@ public class FreeBalanceHarvester {
 					if(!date.trim().equals(""))
 						acf.setTransactionDate(DateTimeUtil.parseDate(date));
 					
-					acf.setTransactionAmount(new BigDecimal(n.doubleValue()));
+					acf.setTransactionAmount(new Double(n.doubleValue()));
 					//ac.getComponentFundings().add(acf);
 				}
