+++ WEB-INF/src/org/digijava/module/aim/dbentity/AmpFundingMTEFProjection.java
@@ -1,6 +1,5 @@
 package org.digijava.module.aim.dbentity;
 
-import java.math.BigDecimal;
 import java.util.Date;
 
 import org.digijava.module.aim.util.FeaturesUtil;
@@ -10,15 +9,15 @@ public class AmpFundingMTEFProjection {
 
 		private Long ampFundingMTEFProjectionId ;
 		private AmpCategoryValue projected; 
-		private BigDecimal amount;
+		private Double amount;
 		private AmpCurrency ampCurrency;
 		private Date projectionDate;
 		private AmpFunding  ampFunding;
 		
-		public BigDecimal getAmount() {
+		public Double getAmount() {
 			return FeaturesUtil.applyThousandsForVisibility(amount);
 		}
-		public void setAmount(BigDecimal amount) {
+		public void setAmount(Double amount) {
 			this.amount = FeaturesUtil.applyThousandsForEntry(amount);
 		}
 		public AmpFunding getAmpFunding() {
