+++ WEB-INF/src/org/digijava/module/aim/dbentity/AmpPledge.java
@@ -1,7 +1,6 @@
 package org.digijava.module.aim.dbentity;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Date;
 
 import org.digijava.module.aim.util.FeaturesUtil;
@@ -13,7 +12,7 @@ public class AmpPledge implements Serializable{
 	private static final long serialVersionUID = -2427620222411238018L;
 	private Long ampPledgeId;
 	private Date date;
-	private BigDecimal amount;
+	private Double amount;
 	private AmpCurrency currency;
 	private Integer adjustmentType ;
 	private String program;
@@ -23,10 +22,10 @@ public class AmpPledge implements Serializable{
 	public void setAdjustmentType(Integer adjustmentType) {
 		this.adjustmentType = adjustmentType;
 	}
-	public BigDecimal getAmount() {
+	public Double getAmount() {
 		return FeaturesUtil.applyThousandsForVisibility(amount);
 	}
-	public void setAmount(BigDecimal amount) {
+	public void setAmount(Double amount) {
 		this.amount = FeaturesUtil.applyThousandsForEntry(amount);
 	}
 	public Long getAmpPledgeId() {
