+++ WEB-INF/src/org/digijava/module/aim/dbentity/AmpReportHierarchy.java
@@ -1,4 +1,4 @@
-package org.digijava.module.aim.dbentity;
+package org.digijava.module.aim.dbentity ;
 
 import java.io.Serializable;
 
@@ -9,52 +9,50 @@ import org.digijava.module.categorymanager.dbentity.AmpCategoryValue;
 import org.digijava.module.categorymanager.util.CategoryConstants;
 import org.digijava.module.categorymanager.util.CategoryManagerUtil;
 
-public class AmpReportHierarchy implements Serializable, Comparable {
+
+public class AmpReportHierarchy implements Serializable, Comparable
+{
 	@ColumnLike
 	private AmpColumns column;
 	@Order
 	private Long levelId;
 	@Level
 	private AmpCategoryValue level;
-
+	
 	private static AmpCategoryValue defaultLevel = null;
-
-	public AmpReportHierarchy() {
-		if (defaultLevel == null)
-			defaultLevel = CategoryManagerUtil.getAmpCategoryValueFromDb(CategoryConstants.ACTIVITY_LEVEL_KEY, (long) 0);
-		level = defaultLevel;
+	
+	public AmpReportHierarchy(){
+		if ( defaultLevel==null )
+			defaultLevel=CategoryManagerUtil.getAmpCategoryValueFromDb(CategoryConstants.ACTIVITY_LEVEL_KEY, (long)0);
+		level	= defaultLevel;
 	}
 
 	public AmpColumns getColumn() {
 		return column;
 	}
-
 	public void setColumn(AmpColumns column) {
 		this.column = column;
 	}
-
 	public Long getLevelId() {
 		return levelId;
 	}
-
 	public void setLevelId(Long levelId) {
 		this.levelId = levelId;
 	}
-
 	public AmpCategoryValue getLevel() {
 		return level;
 	}
-
 	public void setLevel(AmpCategoryValue level) {
 		this.level = level;
 	}
-
+	
 	public int compareTo(Object o) {
 		try {
-			int myOrder = levelId.intValue();
-			int oOrder = ((AmpReportHierarchy) o).getLevelId().intValue();
-			return myOrder - oOrder;
-		} catch (NumberFormatException e) {
+			Long myOrder	= levelId;
+			Long oOrder	=  ((AmpReportHierarchy)o).getLevelId() ;
+			return myOrder.intValue()-oOrder.intValue();
+		}
+		catch (NumberFormatException e) {
 			e.printStackTrace();
 			return -1;
 		}
