+++ WEB-INF/src/org/digijava/module/aim/logic/AmountCalculator.java
@@ -1,6 +1,5 @@
 package org.digijava.module.aim.logic;
 
-import java.math.BigDecimal;
 import java.util.Set;
 
 import org.dgfoundation.amp.ar.cell.CategAmountCell;
@@ -11,6 +10,6 @@ import org.dgfoundation.amp.ar.cell.CategAmountCell;
  */
 public interface AmountCalculator {
 
-	BigDecimal calculateAmount(Set<CategAmountCell> mergedCells);
+	double calculateAmount(Set<CategAmountCell> mergedCells);
 
 }
